setting, contact us, about us, payment partner, footer

+++ tt_website_rodextrip/views/tt_website_rodextrip_views.py
@@ -571,7 +571,7 @@ def admin(request):
                             pass
                     #delete file ga pake
                     for file in os.listdir(fs.location):
-                        if not file in temp and file != 'image_dynamic':
+                        if not file in temp and file != 'image_dynamic' and file != 'image_payment_partner' and file != 'image_about_us':
                             os.remove(fs.location+'/'+file)
             except Exception as e:
                 _logger.error(str(e) + '\n' + traceback.format_exc())
@@ -1044,6 +1044,92 @@ We build this application for our existing partner and public users who register
         'provider_divider_end': math.ceil(len(provider_type) / 2),
         'contact_us': contact_us.split('\n')
     }
+
+
+def contact_us(request):
+    data = []
+    try:
+        file = open(var_log_path() + "contact_data.txt", "r")
+        for line in file:
+            if line != '\n':
+                data.append(line.split('~'))
+        file.close()
+    except:
+        pass
+    # values = {
+    #     'data': data
+    # }
+    # return render(request, MODEL_NAME+'/contact_us.html', values)
+
+    if 'user_account' in request.session._session:
+        try:
+            javascript_version = get_javascript_version()
+            cache_version = get_cache_version()
+            response = get_cache_data(cache_version)
+            airline_country = response['result']['response']['airline']['country']
+            phone_code = []
+            for i in airline_country:
+                if i['phone_code'] not in phone_code:
+                    phone_code.append(i['phone_code'])
+            phone_code = sorted(phone_code)
+            values = get_data_template(request)
+
+            if translation.LANGUAGE_SESSION_KEY in request.session:
+                del request.session[translation.LANGUAGE_SESSION_KEY] #get language from browser
+            values.update({
+                'static_path': path_util.get_static_path(MODEL_NAME),
+                # 'balance': request.session['balance']['balance'] + request.session['balance']['credit_limit'],
+                'titles': ['MR', 'MRS', 'MS', 'MSTR', 'MISS'],
+                'countries': airline_country,
+                'phone_code': phone_code,
+                'username': request.session['user_account'],
+                'static_path_url_server': get_url_static_path(),
+                'javascript_version': javascript_version,
+                'signature': request.session['signature'],
+                'data': data,
+            })
+        except Exception as e:
+            _logger.error(str(e) + '\n' + traceback.format_exc())
+            raise Exception('Make response code 500!')
+        return render(request, MODEL_NAME+'/contact_us.html', values)
+    else:
+        return no_session_logout(request)
+
+
+def about_us(request):
+    if 'user_account' in request.session._session:
+        try:
+            javascript_version = get_javascript_version()
+            cache_version = get_cache_version()
+            response = get_cache_data(cache_version)
+            airline_country = response['result']['response']['airline']['country']
+            phone_code = []
+            for i in airline_country:
+                if i['phone_code'] not in phone_code:
+                    phone_code.append(i['phone_code'])
+            phone_code = sorted(phone_code)
+            values = get_data_template(request)
+
+            if translation.LANGUAGE_SESSION_KEY in request.session:
+                del request.session[translation.LANGUAGE_SESSION_KEY] #get language from browser
+            values.update({
+                'static_path': path_util.get_static_path(MODEL_NAME),
+                # 'balance': request.session['balance']['balance'] + request.session['balance']['credit_limit'],
+                'titles': ['MR', 'MRS', 'MS', 'MSTR', 'MISS'],
+                'countries': airline_country,
+                'phone_code': phone_code,
+                'username': request.session['user_account'],
+                'static_path_url_server': get_url_static_path(),
+                'javascript_version': javascript_version,
+                'signature': request.session['signature'],
+            })
+        except Exception as e:
+            _logger.error(str(e) + '\n' + traceback.format_exc())
+            raise Exception('Make response code 500!')
+        return render(request, MODEL_NAME+'/about_us.html', values)
+    else:
+        return no_session_logout(request)
+
 
 # @api_view(['GET'])
 # def testing(request):
